Add ProgressBar

+++ app/src/main/java/com/zairussalamdev/gitbox/data/entities/UserDetail.kt
@@ -3,30 +3,30 @@ package com.zairussalamdev.gitbox.data.entities
 import com.google.gson.annotations.SerializedName
 
 data class UserDetail(
-    @SerializedName("login")
-    val username: String? = "",
+        @SerializedName("login")
+        val username: String? = "",
 
-    @SerializedName("avatar_url")
-    val avatar: String? = "",
+        @SerializedName("avatar_url")
+        val avatar: String? = "",
 
-    @SerializedName("name")
-    val name: String? = "",
+        @SerializedName("name")
+        val name: String? = "",
 
-    @SerializedName("company")
-    val company: String? = "",
+        @SerializedName("company")
+        val company: String? = "",
 
-    @SerializedName("location")
-    val location: String? = "",
+        @SerializedName("location")
+        val location: String? = "",
 
-    @SerializedName("bio")
-    val bio: String? = "",
+        @SerializedName("bio")
+        val bio: String? = "",
 
-    @SerializedName("public_repos")
-    val repository: Int? = 0,
+        @SerializedName("public_repos")
+        val repository: Int? = 0,
 
-    @SerializedName("followers")
-    val followers: Int? = 0,
+        @SerializedName("followers")
+        val followers: Long? = 0,
 
-    @SerializedName("following")
-    val following: Int? = 0,
+        @SerializedName("following")
+        val following: Long? = 0,
 )
